fix(positions): Add empty state to positions page (#6249)
* fix: Add empty state to positions page

* fix: Add feature flag for positions
