+++ apps/web/src/components/balances/TotalAssetValue/index.tsx
@@ -15,7 +15,7 @@ const TotalAssetValue = ({ fiatTotal }: { fiatTotal: number | undefined }) => {
       </Typography>
       <Typography component="div" variant="h1" fontSize="44px" lineHeight="1.2" letterSpacing="-0.5px">
         {safe.deployed ? (
-          fiatTotal ? (
+          fiatTotal !== undefined ? (
             <FiatValue value={fiatTotal} />
           ) : (
             <Skeleton variant="text" width={60} />
