+++ packages/utils/src/utils/chains.ts
@@ -43,6 +43,7 @@ export enum FEATURES {
   SPACES = 'SPACES',
   EARN = 'EARN',
   MIXPANEL = 'MIXPANEL',
+  POSITIONS = 'POSITIONS',
 }
 
 const MIN_SAFE_VERSION = '1.3.0'
